+++ pycep/parser.py
@@ -1,23 +1,23 @@
 """The pycep python parsing function library."""
 # coding=utf-8
 import re
-import tarfile
+import random
+import string
 
-from sys import exit
 from logging import error, info
+
 from pycep.ceps import CEPS
+from pycep.render import extract_tar_file
 
 
-def extract_tar_file(file_name: str) -> str:
-    """Extract tar gz file from input file_name string."""
-    try:
-        tar_file = tarfile.open(file_name, mode="r:gz")
-        json_file = tar_file.extractfile("package_export.json")
-        return json_file.read()
-    except tarfile.ReadError:
-        error("Input file " + file_name + " is not a proper tar.gz exiting now\n"
-                                          "please correct the file_type input.")
-        exit(1)
+def open_input_file(input_file, file_type):
+    """Return raw data from input file string."""
+    if file_type is "tar":
+        input_data = extract_tar_file(input_file)
+    if file_type == "json":
+        with open(input_file, 'rb') as raw_json:
+            input_data = raw_json.read()
+    return input_data
 
 
 def get_value(item: str, json_package: dict) -> dict:
@@ -30,15 +30,36 @@ def get_value(item: str, json_package: dict) -> dict:
 
 def cep_check_message(cep_number: str):
     """Return cep help message."""
-    return "CEP " + cep_number + " Test Failed!" + " | " + \
-        "More info: https://simspace.github.io/cep/ceps/" + cep_number + "/#requirements"
+    return f"CEP {cep_number} Test Failed! | More info:  https://simspace.github.io/cep/ceps/{cep_number}/#requirements"
 
 
 def get_slide_data(package_export_content_modules: dict, values: str):
     """Return raw data from package and package name."""
-    package_value = package_export_content_modules[values]['contentModuleExportContentModule']
+    package_data = package_export_content_modules[values]
+    package_value = package_data['contentModuleExportContentModule']
     package_name = package_value['name']
     raw_data = ""
+    content_data_node = package_data['contentModuleExportQuestionDescriptions']
+    info(package_name + ": Rendering " + str(len(content_data_node)) + " slides into raw data.")
+    for slide_item in content_data_node:
+        if 'data' in content_data_node[slide_item]:
+            check_dic = content_data_node[slide_item]['data']['document']['nodes']
+            render_slide_data = render_slide(check_dic)
+            if render_slide_data:
+                raw_data += add_newline("# " + package_data['contentModuleExportQuestions'][slide_item]['title']) + \
+                            render_slide_data
+    return raw_data, package_name
+
+
+def h_one_format(heading_data):
+    return f"# {add_newline(heading_data)}"
+
+
+def get_slide_data_listed(package_export_content_modules: dict, values: str):
+    """Return raw data from package and package name."""
+    package_value = package_export_content_modules[values]['contentModuleExportContentModule']
+    package_name = package_value['name']
+    raw_data_dict = {package_name: {}}
     content_data_node = package_export_content_modules[values]['contentModuleExportQuestionDescriptions']
     info(package_name + ": Rendering " + str(len(content_data_node)) + " slides into raw data.")
     for slide_item in content_data_node:
@@ -46,10 +67,13 @@ def get_slide_data(package_export_content_modules: dict, values: str):
             check_dic = content_data_node[slide_item]['data']['document']['nodes']
             render_slide_data = render_slide(check_dic)
             if render_slide_data:
-                raw_data += "# " + package_export_content_modules[values]['contentModuleExportQuestions'][
-                    slide_item]['title'] + "\n"
-                raw_data += render_slide_data
-    return raw_data, package_name
+                raw_data_dict[package_name][package_export_content_modules[values]['contentModuleExportQuestions'][
+                    slide_item]['title']] = render_slide_data
+    return raw_data_dict
+
+
+def id_generator(size=6, chars=string.ascii_lowercase):
+    return ''.join(random.choice(chars) for _ in range(size))
 
 
 def cep_check(raw_slide_data: str, package_name: str) -> None:
@@ -65,15 +89,15 @@ def cep_check(raw_slide_data: str, package_name: str) -> None:
         cep_check_result = cep_check_message(cep_item_dict[cep_items]["cep_number"])
         if cep_item_dict[cep_items]["cep_check_type"] is "contentModuleExportQuestionDescriptions":
             if "string_search" in cep_item_dict[cep_items]:
-                cep_check_result = package_name + ": CEP " + cep_item_dict[cep_items]["cep_number"] + " - Passed"
+                cep_check_result = f"{package_name}: CEP {cep_item_dict[cep_items]['cep_number']} - Passed"
             if "regex_search" in cep_item_dict[cep_items]:
                 test_search = cep_item_dict[cep_items]["regex_search"]
                 x = re.findall(test_search, raw_slide_data)
                 if len(x) > 0:
-                    cep_check_result = package_name + ": CEP " + cep_item_dict[cep_items]["cep_number"] + " - Passed"
+                    cep_check_result = f"{package_name}: CEP {cep_item_dict[cep_items]['cep_number']} - Passed"
             if cep_check_result:
                 if "Passed" not in cep_check_result:
-                    error(package_name + ": " + cep_check_result)
+                    error(f"{package_name}: {cep_check_result}")
                 else:
                     info(cep_check_result)
 
@@ -82,7 +106,9 @@ def render_list_item(slide_line: dict, heading_level: str):
     """Return formatted list data."""
     raw_list_data = ""
     for node in slide_line:
-        raw_list_data += heading_level + node['text']
+        raw_list_data += f"{heading_level}{node['text']}"
+    if len(heading_level) == len(raw_list_data):
+        return None
     return raw_list_data
 
 
@@ -91,14 +117,14 @@ def render_nested_list_nodes(slide_line: dict, format_string: str):
     raw_list_data = ""
     for node in slide_line:
         if 'text' in node:
-            raw_list_data += "   * " + node['text'] + "\n"
+            raw_list_data += add_newline(f"   * {node['text']}")
         else:
             for nested_node in node["nodes"]:
                 if 'text' in nested_node:
-                    raw_list_data += format_string + nested_node['text'] + "\n"
+                    raw_list_data += add_newline(format_string + nested_node['text'])
                 else:
                     if 'text' in nested_node["nodes"][0]:
-                        raw_list_data += "   * " + nested_node["nodes"][0]['text'] + "\n"
+                        raw_list_data += add_newline(f"   * {nested_node['nodes'][0]['text']}")
     return raw_list_data
 
 
@@ -107,6 +133,10 @@ def strip_end_space(string_value: str):
     while string_value[-1:] == " ":
         string_value = string_value[:-1]
     return string_value
+
+
+def add_newline(input_item: str) -> str:
+    return f"{input_item}\n"
 
 
 def render_slide(slide_dict: dict):
@@ -119,28 +149,43 @@ def render_slide(slide_dict: dict):
                     if "marks" in slide_line['nodes'][0] and len(slide_line['nodes'][0]['marks']) > 0:
                         if "bold" == slide_line['nodes'][0]['marks'][0]['type']:
                             if len(slide_line['nodes'][0]['text']) > 0:
-                                raw_slide_data += "***" + strip_end_space(slide_line['nodes'][0]['text']) + "***" + "\n"
+                                raw_slide_data += add_newline(strip_end_space(slide_line['nodes'][0]['text']))
                     else:
-                        raw_slide_data += slide_line['nodes'][0]['text'] + "\n"
+                        raw_slide_data += add_newline(slide_line['nodes'][0]['text'])
                 elif 'unordered-list' in slide_line['type']:
                     raw_list_data = render_nested_list_nodes(slide_line['nodes'], "   * ")
                     if len(raw_list_data) > 0:
                         raw_slide_data += raw_list_data
+                elif 'ordered-list' in slide_line['type']:
+                    raw_list_data = render_nested_list_nodes(slide_line['nodes'], "   * ")
+                    if len(raw_list_data) > 0:
+                        raw_slide_data += add_newline(raw_list_data)
+                elif 'list-item-child' in slide_line['type']:
+                    raw_list_data = render_nested_list_nodes(slide_line['nodes'], "   * ")
+                    if len(raw_list_data) > 0:
+                        raw_slide_data += raw_list_data
                 elif 'heading-two' in slide_line['type']:
-                    raw_heading_data = render_list_item(slide_line['nodes'], '## ')
-                    if len(raw_heading_data) > 3:
-                        raw_slide_data += raw_heading_data + "\n"
+                    raw_heading_data = render_list_item(slide_line['nodes'], '### ')
+                    if raw_heading_data:
+                        raw_slide_data += add_newline(raw_heading_data)
                 elif 'heading-one' in slide_line['type']:
-                    raw_heading_data = render_list_item(slide_line['nodes'], '# ')
-                    if len(raw_heading_data) > 2:
-                        raw_slide_data += raw_heading_data + "\n"
+                    raw_heading_data = render_list_item(slide_line['nodes'], '## ')
+                    if raw_heading_data:
+                        raw_slide_data += add_newline(raw_heading_data)
                 elif 'image-block' in slide_line['type']:
-                    raw_slide_data += "-   []() \n"
+                    picture_id = id_generator()
+                    raw_heading_data = f"![{picture_id}]({slide_line['data']['imageData']})"
+                    if raw_heading_data:
+                        raw_slide_data += add_newline(raw_heading_data)
                 elif 'code-block' in slide_line['type']:
                     raw_heading_data = render_nested_list_nodes(slide_line['nodes'], "    ")
-                    if len(raw_heading_data) > 3:
+                    if raw_heading_data:
                         raw_slide_data += raw_heading_data
+                elif 'code-line' in slide_line['type']:
+                    raw_heading_data = render_nested_list_nodes(slide_line['nodes'], "    ")
+                    if raw_heading_data and len(raw_heading_data) > 5:
+                        raw_slide_data += add_newline(raw_heading_data)
                 else:
                     error("Data type " + slide_line['type'] + "unknown")
-                    raw_slide_data += "Data Type Unknown \n"
+                    raw_slide_data += " Data Type Unknown\n"
     return raw_slide_data