+++ CHANGELOG.md
@@ -8,3 +8,14 @@
 -   Python documentation style fixes
 -   Moved plugin to dedicated pycep library 
 -   Docker documentation added
+
+## pycep Release 0.0.3
+-   Spellcheck plugin added
+-   Markdown output rendering support added
+-   Add user word list feature
+-   Create render python library to handle data output
+-   Create MANIFEST.in to include data in library install
+-   Add pyspellcheck to requirements.txt
+-   Move code of conduct to docs folder
+-   Update arg help functions
+-   Add Contributing.md to docs